fix(harness): run live e2e on PRs and hide blank-slate PR stats
Interchange sticky chrome always mounted the committed-PR band, which
broke zero-repo e2e and blank-slate UX. Gate the band after membership
loads empty, point Gherkin at Jobs/Pipeline nav, extract the shared
repositories query, and run vault-backed live e2e on same-repo PRs
(skip when the secret is unavailable for forks).
