+++ docs/adr/0021-live-harness-end-to-end-test.md
@@ -12,6 +12,6 @@ Throwaway project on `git.drupalcode.org` under operator control (default Projec
 
 ## Shared run model
 
-Each run starts with an empty, isolated Harness database. CI copies a checked-in encrypted Keymaxxer vault into a temporary home and unlocks it with the GitHub Actions secret `E2E_KEYMAXXER_MASTER_KEY`; `KEYMAXXER_APPROVE=deny` prevents headless prompts and fails closed. Local runs instead leave `~/.keymaxxer` untouched and may prompt normally. Scenarios have no automatic retry, run only after the existing checks on trusted pushes to `main`, and retain Playwright diagnostics on failure. Scenarios assert the sentinel's fixed identity and do not reject unrelated Issues on the fixture project.
+Each run starts with an empty, isolated Harness database. CI copies a checked-in encrypted Keymaxxer vault into a temporary home and unlocks it with the GitHub Actions secret `E2E_KEYMAXXER_MASTER_KEY`; `KEYMAXXER_APPROVE=deny` prevents headless prompts and fails closed. Local runs instead leave `~/.keymaxxer` untouched and may prompt normally. Scenarios have no automatic retry. Trusted `push` to `main` fail-closes if the vault secret is missing; pull requests run live e2e when the secret is available (same-repo) and skip it when unavailable (fork PRs), while still running quality gates. Playwright diagnostics are retained on failure. Scenarios assert the sentinel's fixed identity and do not reject unrelated Issues on the fixture project.
 
 Work Item creation, Agent Turns, pull request / merge request creation, status checks, review, merge, and cleanup remain out of scope for the automated live e2e (operator demos may exercise the full GitLab lifecycle separately with write credentials).
