+++ docs/adr/0024-manual-oidc-npm-release.md
@@ -1,3 +1,3 @@
 # Manual conventional-commit release with npm OIDC
 
-Releases are cut only by **manual `workflow_dispatch`** on `.github/workflows/ci-cd.yml` (**CI/CD**), not on every green `main` push. Version is computed from **conventional commits** since the last tag (first public line `0.1.0`; breaking changes bump major even on 0.x). If there is nothing to release, the workflow **fails**. Every `push` to `main` (and every release dispatch) runs a host **packed-install** smoke job that packs the launcher and host platform package, installs them outside the checkout, and exercises the installed command without Bun/Nx; the release job **depends on that job** and re-runs the same smoke against versioned release binaries before publish. Cross-compiled platform packages keep structural validation only. The release job then publishes launcher + platform packages to npm with **Trusted Publishing (OIDC)** and **provenance** (no long-lived `NPM_TOKEN`), tags `vX.Y.Z`, and creates a **GitHub Release** with notes and the same binaries. Pull requests use `.github/workflows/pr.yml` (**PR**); `push` to `main` uses CI/CD quality gates (lint, knip, test, typecheck), the packed-install gate, and the live Harness e2e step (vault key never available to PR workflows). Live Gherkin e2e remains complementary and does not replace the package-boundary test. Rejected: continuous publish from every merge, free-form version input without commit analysis, and npm token secrets in Actions.
+Releases are cut only by **manual `workflow_dispatch`** on `.github/workflows/ci-cd.yml` (**CI/CD**), not on every green `main` push. Version is computed from **conventional commits** since the last tag (first public line `0.1.0`; breaking changes bump major even on 0.x). If there is nothing to release, the workflow **fails**. Every `push` to `main` (and every release dispatch) runs a host **packed-install** smoke job that packs the launcher and host platform package, installs them outside the checkout, and exercises the installed command without Bun/Nx; the release job **depends on that job** and re-runs the same smoke against versioned release binaries before publish. Cross-compiled platform packages keep structural validation only. The release job then publishes launcher + platform packages to npm with **Trusted Publishing (OIDC)** and **provenance** (no long-lived `NPM_TOKEN`), tags `vX.Y.Z`, and creates a **GitHub Release** with notes and the same binaries. Pull requests use `.github/workflows/pr.yml` (**PR**); `push` to `main` uses CI/CD quality gates (lint, knip, test, typecheck), the packed-install gate, and the live Harness e2e step. Main fails closed without `E2E_KEYMAXXER_MASTER_KEY`; PRs run live Gherkin e2e when that secret is available (same-repo) and skip it for fork PRs. Live Gherkin e2e remains complementary and does not replace the package-boundary test. Rejected: continuous publish from every merge, free-form version input without commit analysis, and npm token secrets in Actions.
